+++ src/Grant/AuthCodeGrant.php
@@ -343,7 +343,7 @@ class AuthCodeGrant extends AbstractAuthorizeGrant
                 'client_id'             => $authCode->getClient()->getIdentifier(),
                 'redirect_uri'          => $authCode->getRedirectUri(),
                 'auth_code_id'          => $authCode->getIdentifier(),
-                'scopes'                => $authCode->getScopes(),
+                'scopes'                => $authCode->getScopeIdentifiers(),
                 'user_id'               => $authCode->getUserIdentifier(),
                 'expire_time'           => (new \DateTime())->add($this->authCodeTTL)->format('U'),
                 'code_challenge'        => $authorizationRequest->getCodeChallenge(),